+++ src/libOpenImageIO/color_ocio.cpp
@@ -464,7 +464,7 @@ ColorConfig::Impl::test_conversion_yields(const char* from, const char* to,
     if (!proc)
         return false;
     OIIO_DASSERT(test_colors.size() == result_colors.size());
-    int n              = std::ssize(test_colors);
+    int n              = (int)std::ssize(test_colors);
     Imath::C3f* colors = OIIO_ALLOCA(Imath::C3f, n);
     std::copy(test_colors.data(), test_colors.data() + n, colors);
     proc->apply((float*)colors, n, 1, 3, sizeof(float), 3 * sizeof(float),
@@ -845,7 +845,7 @@ ColorConfig::getAliases(string_view color_space) const
     if (config) {
         auto cs = config->getColorSpace(c_str(color_space));
         if (cs) {
-            for (int i = 0, e = cs->getNumAliases(); i < e; ++i)
+            for (size_t i = 0, e = cs->getNumAliases(); i < e; ++i)
                 result.emplace_back(cs->getAlias(i));
         }
     }
